refactor some gplately.tools into new spatial module
